feat: add links to different map services
Apple Maps links used to open Google Maps on devices without the Apple Maps app.
But now that Apple opened the web app version of their map service
it doesn't open Google Maps any more.
That's not great for users of Google Maps, so now we give the option explicitly in the UI.

And we took that opportunity to add a link to OpenStreetMap.